fix screenreader issue for usernames in messages
