+++ utils/post_utils.ts
@@ -12,10 +12,10 @@ import {Permissions, Posts} from 'mattermost-redux/constants';
 import {getChannel} from 'mattermost-redux/selectors/entities/channels';
 import {getConfig} from 'mattermost-redux/selectors/entities/general';
 import {makeGetReactionsForPost} from 'mattermost-redux/selectors/entities/posts';
-import {get, isCollapsedThreadsEnabled} from 'mattermost-redux/selectors/entities/preferences';
+import {get, getTeammateNameDisplaySetting, isCollapsedThreadsEnabled} from 'mattermost-redux/selectors/entities/preferences';
 import {haveIChannelPermission} from 'mattermost-redux/selectors/entities/roles';
 import {getCurrentTeamId, getTeam} from 'mattermost-redux/selectors/entities/teams';
-import {makeGetDisplayName, getCurrentUserId, getUser, UserMentionKey} from 'mattermost-redux/selectors/entities/users';
+import {makeGetDisplayName, getCurrentUserId, getUser, UserMentionKey, getUsersByUsername} from 'mattermost-redux/selectors/entities/users';
 
 import {Channel} from '@mattermost/types/channels';
 import {ClientConfig, ClientLicense} from '@mattermost/types/config';
@@ -23,10 +23,12 @@ import {ServerError} from '@mattermost/types/errors';
 import {Group} from '@mattermost/types/groups';
 import {Post} from '@mattermost/types/posts';
 import {Reaction} from '@mattermost/types/reactions';
+import {UserProfile} from '@mattermost/types/users';
 
 import {getUserIdFromChannelName} from 'mattermost-redux/utils/channel_utils';
 import * as PostListUtils from 'mattermost-redux/utils/post_list';
 import {canEditPost as canEditPostRedux} from 'mattermost-redux/utils/post_utils';
+import {displayUsername} from 'mattermost-redux/utils/user_utils';
 
 import {getEmojiMap} from 'selectors/emojis';
 import {getIsMobileView} from 'selectors/views/browser';
@@ -426,13 +428,15 @@ export function makeCreateAriaLabelForPost(): (state: GlobalState, post: Post) =
         (state: GlobalState, post: Post) => getReactionsForPost(state, post.id),
         (state: GlobalState, post: Post) => get(state, Preferences.CATEGORY_FLAGGED_POST, post.id, null) != null,
         getEmojiMap,
-        (post, author, reactions, isFlagged, emojiMap) => {
-            return (intl: IntlShape) => createAriaLabelForPost(post, author, isFlagged, reactions ?? {}, intl, emojiMap);
+        (state: GlobalState) => getUsersByUsername(state),
+        (state: GlobalState) => getTeammateNameDisplaySetting(state),
+        (post, author, reactions, isFlagged, emojiMap, users, teammateNameDisplaySetting) => {
+            return (intl: IntlShape) => createAriaLabelForPost(post, author, isFlagged, reactions ?? {}, intl, emojiMap, users, teammateNameDisplaySetting);
         },
     );
 }
 
-export function createAriaLabelForPost(post: Post, author: string, isFlagged: boolean, reactions: Record<string, Reaction>, intl: IntlShape, emojiMap: EmojiMap): string {
+export function createAriaLabelForPost(post: Post, author: string, isFlagged: boolean, reactions: Record<string, Reaction>, intl: IntlShape, emojiMap: EmojiMap, users: Record<string, UserProfile>, teammateNameDisplaySetting: string): string {
     const {formatMessage, formatTime, formatDate} = intl;
 
     let message = post.state === Posts.POST_DELETED ? formatMessage({
@@ -450,6 +454,17 @@ export function createAriaLabelForPost(post: Post, author: string, isFlagged: bo
     while ((match = Emoticons.EMOJI_PATTERN.exec(message)) !== null) {
         if (emojiMap.has(match[2])) {
             message = message.replace(match[0], `${match[2].replace(/_/g, ' ')} emoji`);
+        }
+    }
+
+    // Replace mentions with preferred username
+    const mentionsArray = message.match(Constants.MENTIONS_WITHOUT_SPECIAL_MENTIONS) || [];
+    for (let i = 0; i < mentionsArray.length; i++) {
+        const mention = mentionsArray[i];
+        const user  = getUserFromMentionName(users, mention.substring(1));
+
+        if (user) {
+            message = message.replace(mention, `@${displayUsername(user, teammateNameDisplaySetting)}`);
         }
     }
 
@@ -640,4 +655,24 @@ export function makeGetUniqueReactionsToPost(): (state: GlobalState, postId: Pos
             return reactionsForPost;
         },
     );
+}
+
+
+export function getUserFromMentionName(usersByUsername: Record<string, UserProfile>, mentionName: string) {
+    let mentionNameToLowerCase = mentionName.toLowerCase();
+
+    while (mentionNameToLowerCase.length > 0) {
+        if (usersByUsername.hasOwnProperty(mentionNameToLowerCase)) {
+            return usersByUsername[mentionNameToLowerCase];
+        }
+
+        // Repeatedly trim off trailing punctuation in case this is at the end of a sentence
+        if ((/[._-]$/).test(mentionNameToLowerCase)) {
+            mentionNameToLowerCase = mentionNameToLowerCase.substring(0, mentionNameToLowerCase.length - 1);
+        } else {
+            break;
+        }
+    }
+
+    return '';
 }
